Use a generated CHANGES.md for installers. This replaces the first occurrence of `## SNAPSHOT` with the full version number into `target/CHANGES.md`. This file is then embedded in the installer.
After we promote, the source CHANGES.md should be changed so that `## SNAPSHOT` is manually changed to the new build number, and a new `## SNAPSHOT` line added near the top of the file (the "version bump"). As @cdakin is the one who promotes versions, and knows what version this happens on, he should be the one to bump the CHANGES.md file.